+++ CHANGELOG.md
@@ -2,6 +2,11 @@ Sponsored by [MyText.ai](https://mytext.ai)
 
 [![](./example/SponsoredByMyTextAi.png)](https://mytext.ai)
 
+## 1.4.0
+
+* UserException now has `fromJson`, `toJson`, and `copyWith` methods.
+  This makes it compatible with Serverpod.
+
 ## 1.3.1
 
 * Fixed translation typo bug.
@@ -16,6 +21,8 @@ Sponsored by [MyText.ai](https://mytext.ai)
 
 ## 1.0.5
 
-* On March 5, 2024, I've created this Dart-only package to contain some Dart-only code used by
-  the [async_redux](https://pub.dev/packages/async_redux) package. For the moment, this Dart-only
+* On March 5, 2024, I've created this Dart-only package to contain some
+  Dart-only code used by
+  the [async_redux](https://pub.dev/packages/async_redux) package. For the
+  moment, this Dart-only
   package simply contains the `UserException`, and nothing more.